Create Timeseria anomaly detector files and folders
