+++ ats/anomaly_detectors/dl/LSTM_anomaly_detector.py
@@ -1,0 +1,5 @@
+from ..base import TimeseriaAnomalyDetector
+from timeseria.models.anomaly_detectors import LSTMAnomalyDetector as TimeseriaLSTMAnomalyDetector  
+
+class LSTMAnomalyDetector(TimeseriaAnomalyDetector):
+    model_class = TimeseriaLSTMAnomalyDetector
