+++ ats/anomaly_detectors/ml/linear_regression_anomaly_detector.py
@@ -1,0 +1,5 @@
+from ..base import TimeseriaAnomalyDetector
+from timeseria.models.anomaly_detectors import LinearRegressionAnomalyDetector as TimeseriaLinearRegressionAnomalyDetector              
+
+class LinearRegressionAnomalyDetector(TimeseriaAnomalyDetector):
+    model_class = TimeseriaLinearRegressionAnomalyDetector
